frontend/requirements.txt
google-generativeai
python-dotenv
flask
pdf2image
markdown
numpy
pandas
seaborn
matplotlib
scikit-learn